+++ src/ModularPipelines/Engine/ModuleIgnoreHandler.cs
@@ -18,9 +18,22 @@ internal class ModuleIgnoreHandler : IModuleIgnoreHandler
 
     public async Task<bool> ShouldIgnore(ModuleBase module)
     {
-        if (IsIgnoreCategory(module) || !IsRunnableCategory(module) || !await IsRunnableCondition(module))
+        if (IsIgnoreCategory(module))
         {
-            module.SetSkipped();
+            module.SetSkipped("A category of this module has been ignored");
+            return true;
+        }
+        
+        if (!IsRunnableCategory(module))
+        {
+            module.SetSkipped("The module was not in a runnable category");
+            return true;
+        }
+
+        var (isRunnableCondition, runConditionAttribute) = await IsRunnableCondition(module);
+        if (!isRunnableCondition)
+        {
+            module.SetSkipped($"A condition to run this module has not been met - {runConditionAttribute}");
             return true;
         }
 
@@ -55,29 +68,35 @@ internal class ModuleIgnoreHandler : IModuleIgnoreHandler
         return category != null && ignoreCategories.Contains(category.Category);
     }
 
-    private async Task<bool> IsRunnableCondition(ModuleBase module)
+    private async Task<(bool, RunConditionAttribute?)> IsRunnableCondition(ModuleBase module)
     {
         var mandatoryRunConditionAttributes = module.GetType().GetCustomAttributes<MandatoryRunConditionAttribute>(true).ToList();
         var runConditionAttributes = module.GetType().GetCustomAttributes<RunConditionAttribute>(true).Except(mandatoryRunConditionAttributes).ToList();
 
         var mandatoryConditionResults = await mandatoryRunConditionAttributes.ToAsyncProcessorBuilder()
-            .SelectAsync(async runConditionAttribute => await runConditionAttribute.Condition(module.Context))
+            .SelectAsync(async runConditionAttribute => new RunnableConditionMet(await runConditionAttribute.Condition(module.Context), runConditionAttribute))
             .ProcessInParallel();
 
-        if (mandatoryConditionResults.Any(result => !result))
+        var mandatoryCondition = mandatoryConditionResults.FirstOrDefault(result => !result.ConditionMet);
+        
+        if (mandatoryCondition != null)
         {
-            return false;
+            return (false, mandatoryCondition.RunConditionAttribute);
         }
         
         if (!runConditionAttributes.Any())
         {
-            return true;
+            return (true, null);
         }
         
         var conditionResults = await runConditionAttributes.ToAsyncProcessorBuilder()
-            .SelectAsync(async runConditionAttribute => await runConditionAttribute.Condition(module.Context))
+            .SelectAsync(async runConditionAttribute => new RunnableConditionMet(await runConditionAttribute.Condition(module.Context), runConditionAttribute))
             .ProcessInParallel();
 
-        return conditionResults.Any(result => result);
+        var runnableCondition = conditionResults.FirstOrDefault(result => result.ConditionMet);
+        
+        return (runnableCondition != null, null);
     }
+
+    private record RunnableConditionMet(bool ConditionMet, RunConditionAttribute RunConditionAttribute);
 }