+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
-*   Refactor command logging with bit flag enums
+* Breaking: ShouldSkip virtual method now returns a `SkipDecision` class that allows you to plug in a reason
+* Breaking: Refactor command logging with bit flag enums